moved shared components into Shared folder
