+++ client/src/routes/Transaction.js
@@ -1,16 +1,16 @@
 import React, { useState } from "react"
-import Header from "../components/Header"
-import Footer from "../components/Footer"
-import SiteWrapper from "../components/SiteWrapper"
+import Header from "../components/Shared/Header"
+import Footer from "../components/Shared/Footer"
+import SiteWrapper from "../components/Shared/SiteWrapper"
 import { useExpenses } from "../contexts/Expenses"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faArrowDown } from "@fortawesome/free-solid-svg-icons"
 import createTransaction from "../services/createTransaction"
 import getExpenses from "../services/getExpenses"
 import { useNavigate } from "react-router-dom"
-import Button from "../components/Button"
-import Spinner from "../components/Spinner"
-import Protected from "../components/Protected"
+import Button from "../components/Shared/Button"
+import Spinner from "../components/Shared/Spinner"
+import Protected from "../components/Shared/Protected"
 
 export default function AddTransaction() {
     // UI state
